+++ Changelog.md
@@ -37,6 +37,7 @@
 * Fix problem with parsing absolute file path in Windows 
 * Parse `keep_next` in paragraph properties
 * TransformEffect in correct `OoxmlSize` unit
+* `Indents` use `OoxmlSize`
 
 ### Refactor
 * Redone parsing images - store in structure, instead of copying file to filesystem